+++ crates/ruma-signatures/src/verify/tests.rs
@@ -31,7 +31,7 @@ fn add_key_to_map(public_key_map: &mut PublicKeyMap, name: &str, pair: &Ed25519K
     let encoded_public_key = Base64::new(pair.public_key().to_vec());
     let version = ServerSigningKeyId::from_parts(
         SigningKeyAlgorithm::Ed25519,
-        pair.version().try_into().unwrap(),
+        &pair.version().try_into().unwrap(),
     );
 
     sender_key_map.insert(version.to_string(), encoded_public_key);
@@ -42,7 +42,7 @@ fn add_invalid_key_to_map(public_key_map: &mut PublicKeyMap, name: &str, pair: &
     let encoded_public_key = Base64::new(pair.public_key().to_vec());
     let version = ServerSigningKeyId::from_parts(
         SigningKeyAlgorithm::from("an-unknown-algorithm"),
-        pair.version().try_into().unwrap(),
+        &pair.version().try_into().unwrap(),
     );
 
     sender_key_map.insert(version.to_string(), encoded_public_key);
@@ -273,7 +273,7 @@ fn verify_event_fails_if_public_key_is_invalid() {
     let encoded_public_key = Base64::new(newly_generated_key_pair.public_key().to_vec());
     let version = ServerSigningKeyId::from_parts(
         SigningKeyAlgorithm::Ed25519,
-        key_pair_sender.version().try_into().unwrap(),
+        &key_pair_sender.version().try_into().unwrap(),
     );
     sender_key_map.insert(version.to_string(), encoded_public_key);
     public_key_map.insert("domain-sender".to_owned(), sender_key_map);